+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.0.2
+
+- mobile button press fix
+- start astroid with 1
+- fix caching issues.
+
+
 ## 1.0.1
 
 - work on client server updates